+++ src/assets/data/entries/entries.json
@@ -54,7 +54,7 @@
         "title": "FURTHER RESOURCES",
         "contents": [
           "Beamsplitters are a fundamental optical element in all optics and quantum physics labs. They are commonly used to create more beams from a single incoming beam, such that the two outgoing beams can be recombined later after one of them has undergone further manipulation. Some common setups which embed beamsplitters are:<br/>-Michelson-Morley interferometers<br/>-Mach-Zehnder interferometers<br/>",
-          "encyclopedia/Beamsplitter-1.png",
+          "../assets/encyclopedia/Beamsplitter-1.png",
           "Famous experiments that used beamsplitters are:<br/>-The Fizeau experiment of 1851 to measure the speeds of light in water<br/>-The Michelson-Morley experiment of 1887 to measure the effect of the (hypothetical) luminiferous aether on the speed of light<br/>-Bell test experiments (from ca. 1972) to demonstrate consequences of quantum entanglement and exclude local hidden variable theories<br/>-Wheeler's delayed choice experiment of 1978, 1984 etc., to test what makes a photon behave as a wave or a particle and when it happens"
         ]
       },
