+++ src/components/Board/Cell/AbsorberCell.vue
