+++ src/components/Board/Cell/BeamSplitterCell.vue
