+++ src/components/Board/Cell/CoatedBeamSplitterCell.vue
