+++ src/components/Board/Cell/CornerCubeCell.vue
