+++ src/components/Board/Cell/DetectorCell.vue
