+++ src/components/Board/Cell/DetectorFourCell.vue
