+++ src/components/Board/Cell/FaradayRotatorCell.vue
