+++ src/components/Board/Cell/LaserCell.vue
