+++ src/components/Board/Cell/MineCell.vue
