+++ src/components/Board/Cell/MirrorCell.vue
