+++ src/components/Board/Cell/Piece.ts
