+++ src/components/Board/Cell/PolarizerCell.vue
