+++ src/components/Board/Cell/PolarizingBeamSplitterCell.vue
