+++ src/components/Board/Cell/RockCell.vue
