+++ src/components/Board/Cell/SugarSolutionCell.vue
