+++ src/components/Board/Cell/VacuumJarCell.vue
