+++ src/components/Board/Cell/index.ts
